small fixes to group service hook from PR comments
