+++ client/allocrunner/interfaces/runner_lifecycle.go
@@ -44,7 +44,6 @@ type RunnerUpdateHook interface {
 
 type RunnerUpdateRequest struct {
 	Alloc *structs.Allocation
-	Node  *structs.Node
 }
 
 // ShutdownHook may be implemented by AllocRunner or TaskRunner hooks and will
